// containers/futures/futures-manager-container.tsx
import React, { useMemo, useState } from 'react'
import { parseUnits } from 'viem'
import Link from 'next/link'
import { useRouter } from 'next/router'

import { Asset } from '../../model/futures/asset'
import { MintFuturesAssetForm } from '../../components/form/futures/mint-futures-asset-form'
import { useCurrencyContext } from '../../contexts/currency-context'
import {
  calculateLiquidationPrice,
  calculateLtv,
  calculateMaxLoanableAmount,
} from '../../utils/ltv'
import { formatUnits } from '../../utils/bigint'
import { useFuturesContractContext } from '../../contexts/futures/futures-contract-context'
import Modal from '../../components/modal/modal'
import { useChainContext } from '../../contexts/chain-context'
import { isMarketClose } from '../../utils/date'

export const FuturesManagerContainer = ({ asset }: { asset: Asset }) => {
  const router = useRouter()
  const { selectedChain } = useChainContext()
  const [displayMarketClosedModal, setDisplayMarketClosedModal] =
    useState(false)
  const { balances, prices } = useCurrencyContext()
  const { borrow } = useFuturesContractContext()
  const [collateralValue, setCollateralValue] = useState('')
  const [borrowValue, setBorrowValue] = useState('')

  const [debtAmount, collateralAmount, collateralUserBalance] = useMemo(() => {
    return [
      parseUnits(borrowValue || '0', asset.currency.decimals),
      parseUnits(collateralValue || '0', asset.collateral.decimals),
      balances[asset.collateral.address] ?? 0n,
    ]
  }, [
    asset.collateral.address,
    asset.collateral.decimals,
    asset.currency.decimals,
    balances,
    borrowValue,
    collateralValue,
  ])

  const maxBorrowAmount = useMemo(
    () =>
      calculateMaxLoanableAmount(
        asset.currency,
        parseUnits((prices[asset.currency.address] ?? 0).toFixed(18), 18),
        asset.collateral,
        parseUnits((prices[asset.collateral.address] ?? 0).toFixed(18), 18),
        collateralAmount,
        asset.maxLTV,
        asset.ltvPrecision,
      ),
    [
      asset.collateral,
      asset.currency,
      asset.ltvPrecision,
      asset.maxLTV,
      collateralAmount,
      prices,
    ],
  )
  const isDeptSizeLessThanMinDebtSize = useMemo(
    () => asset.minDebt > 0n && debtAmount < asset.minDebt,
    [asset.minDebt, debtAmount],
  )

  return displayMarketClosedModal ? (
    <Modal
      show
      onClose={() => {}}
      onButtonClick={() => setDisplayMarketClosedModal(false)}
    >
      <h1 className="flex font-bold text-xl mb-2">Notice</h1>
      <div className="text-sm">
        our price feeds follow the traditional market hours of each asset
        classes and will be available at the following hours:{' '}
        <span>
          <Link
            className="text-blue-500 underline font-bold"
            target="_blank"
            href="https://docs.pyth.network/price-feeds/market-hours"
          >
            [Link]
          </Link>
        </span>
      </div>
    </Modal>
  ) : (
    <MintFuturesAssetForm
      chain={selectedChain}
      asset={asset}
      maxBorrowAmount={maxBorrowAmount}
      borrowLTV={calculateLtv(
        asset.currency,
        prices[asset.currency.address] ?? 0,
        debtAmount,
        asset.collateral,
        prices[asset.collateral.address] ?? 0,
        collateralAmount,
      )}
      collateralValue={collateralValue}
      setCollateralValue={setCollateralValue}
      borrowValue={borrowValue}
      setBorrowValue={setBorrowValue}
      balances={balances}
      prices={prices}
      liquidationPrice={calculateLiquidationPrice(
        asset.currency,
        prices[asset.currency.address] ?? 0,
        asset.collateral,
        prices[asset.collateral.address] ?? 0,
        debtAmount,
        collateralAmount,
        asset.liquidationThreshold,
        asset.ltvPrecision,
      )}
      actionButtonProps={{
        disabled:
          collateralAmount === 0n ||
          debtAmount === 0n ||
          collateralAmount > collateralUserBalance ||
          debtAmount > maxBorrowAmount ||
          isDeptSizeLessThanMinDebtSize,
        onClick: async () => {
          const closed = isMarketClose(asset.currency.priceFeedId)
          if (closed) {
            setDisplayMarketClosedModal(true)
            return
          }
          const hash = await borrow(asset, collateralAmount, debtAmount)
          if (hash) {
            await router.replace('/futures')
          }
        },
        text:
          collateralAmount === 0n
            ? 'Enter collateral amount'
            : debtAmount === 0n
              ? 'Enter loan amount'
              : collateralAmount > collateralUserBalance
                ? `Insufficient ${asset.collateral.symbol} balance`
                : debtAmount > maxBorrowAmount
                  ? 'Not enough collateral'
                  : isDeptSizeLessThanMinDebtSize
                    ? `Remaining debt must be ≥ ${formatUnits(
                        asset.minDebt,
                        asset.currency.decimals,
                        prices[asset.currency.address] ?? 0,
                      )} ${asset.currency.symbol.split('-')[0]}`
                    : 'Mint',
      }}
    />
  )
}
